requirements.txt
django
djangorestframework
django-mptt
djoser
python-dotenv
